pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "king-tensor"
version = "0.1.0"
description = "Dominate Clash Royale with AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "mss>=10.1.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pyautogui>=0.9.54",
    "pyyaml>=6.0.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[project.scripts]
kt = "king_tensor.cli:main"

[tool.setuptools]
packages = ["king_tensor"]

[tool.uv.workspace]
members = [
    ".",
    ".",
    ".",
    ".",
]

[tool.uv.sources]
king-tensor = { workspace = true }

[dependency-groups]
dev = [
    "king-tensor",
]
